Move useFetchNfts hook to a separate file
